clk_meinberg: change 4 implicit casts to explicit.

+++ libparse/clk_meinberg.c
@@ -398,7 +398,8 @@ mbg_input(
 {
 	unsigned int rtc;
 
-	parseprintf(DD_PARSE, ("mbg_input(0x%lx, 0x%x, ...)\n", (long)parseio, ch));
+	parseprintf(DD_PARSE, ("mbg_input(0x%lx, 0x%x, ...)\n",
+                    (unsigned long)parseio, (unsigned)ch));
 
 	switch (ch)
 	{
@@ -566,7 +567,8 @@ gps_input(
 
   msg_buf = (struct msg_buf *)parseio->parse_pdata;
 
-  parseprintf(DD_PARSE, ("gps_input(0x%lx, 0x%x, ...)\n", (long)parseio, ch));
+  parseprintf(DD_PARSE, ("gps_input(0x%lx, 0x%x, ...)\n",
+              (unsigned long)parseio, (unsigned)ch));
 
   if (!msg_buf)
     return PARSE_INP_SKIP;
